Move browser sync config into theme
