+++ .env.example
@@ -43,12 +43,6 @@ WP_REDIS_DISABLED=true
 ### Disable WP Cron
 DISABLE_WP_CRON=true
 
-### Webpack BrowserSync
-#BROWSER_SYNC_ENABLE=true
-#BROWSER_SYNC_PROXY=https://sitename.local
-#BROWSER_SYNC_PORT=3003
-#BROWSER_SYNC_HTTPS=true
-
 ### Generate your keys here: https://roots.io/salts.html
 AUTH_KEY='generateme'
 SECURE_AUTH_KEY='generateme'
