+++ package.json
@@ -17,10 +17,7 @@
 		"@wordpress/scripts": "^30.0.2",
 		"@wordpress/stylelint-config": "^23.0.0",
 		"autoprefixer": "^10.4.20",
-		"browser-sync": "^3.0.3",
-		"browser-sync-v3-webpack-plugin": "^0.1.0",
 		"cssnano": "^7.0.6",
-		"dotenv": "^16.4.5",
 		"eslint-plugin-prettier": "^5.2.1",
 		"husky": "^9.1.6",
 		"lint-staged": "^15.2.10",
