+++ packages/themes/block-theme/.env.example
@@ -1,0 +1,5 @@
+### Webpack BrowserSync
+#BROWSER_SYNC_ENABLE=true
+#BROWSER_SYNC_PROXY=https://project-base.local
+#BROWSER_SYNC_PORT=3003
+#BROWSER_SYNC_HTTPS=true
