+++ packages/themes/block-theme/package.json
@@ -4,6 +4,9 @@
 	"version": "1.0.0",
 	"devDependencies": {
 		"@wordpress/scripts": "^30.0.2",
+		"browser-sync": "^3.0.3",
+		"browser-sync-v3-webpack-plugin": "^0.1.0",
+		"dotenv": "^16.4.5",
 		"fast-glob": "^3.3.2"
 	},
 	"scripts": {
