+++ packages/themes/block-theme/webpack.config.js
@@ -3,6 +3,8 @@
  */
 const { sync: globSync } = require('fast-glob');
 const path = require('path');
+const BrowserSyncPlugin = require('browser-sync-v3-webpack-plugin');
+require('dotenv').config();
 
 /**
  * WordPress dependencies
@@ -28,7 +30,32 @@ function getPackageEntryPoints() {
 	return entries;
 }
 
-module.exports = {
+const config = {
 	...rootConfig,
 	entry: getPackageEntryPoints(),
 };
+
+if ('true' === process.env.BROWSER_SYNC_ENABLE) {
+	module.exports = {
+		...config,
+		plugins: [
+			...config.plugins.filter(
+				(plugin) => plugin.constructor.name !== 'BrowserSyncPlugin',
+			),
+			new BrowserSyncPlugin(
+				{
+					files: ['**/*.css', '**/*.js'],
+					proxy:
+						process.env.BROWSER_SYNC_PROXY ?? process.env.WP_HOME,
+					port: process.env.BROWSER_SYNC_PORT ?? 3002,
+					https: 'true' === process.env.BROWSER_SYNC_HTTPS,
+				},
+				{
+					reload: false,
+				},
+			),
+		].filter(Boolean),
+	};
+}
+
+module.exports = config;
